+++ cognite/client/testing.py
@@ -48,7 +48,8 @@ from cognite.client._api.relationships import RelationshipsAPI
 from cognite.client._api.sequences import SequencesAPI, SequencesDataAPI
 from cognite.client._api.simulators import SimulatorsAPI
 from cognite.client._api.simulators.integrations import SimulatorIntegrationsAPI
-from cognite.client._api.simulators.models import SimulatorModelRevisionsAPI, SimulatorModelsAPI
+from cognite.client._api.simulators.models import SimulatorModelsAPI
+from cognite.client._api.simulators.models_revisions import SimulatorModelRevisionsAPI
 from cognite.client._api.synthetic_time_series import SyntheticDatapointsAPI
 from cognite.client._api.templates import (
     TemplateGroupsAPI,
